# web/swagger-converted.yml
openapi: 3.0.0
info:
  title: Neucore API
  description: Client library of Neucore API
  version: 1.0.0
paths:
...
  "/user/alliance/{id}/add-group/{gid}":
    put:
      tags:
        - Alliance
      summary: Add a group to the alliance.
      description: "Needs role: group-admin"
      operationId: addGroup
      parameters:
        - name: id
          in: path
          description: ID of the alliance.
          required: true
          schema:
            type: integer
        - name: gid
          in: path
          description: ID of the group.
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Group added.
        "403":
          description: Not authorized.
        "404":
          description: Alliance and/or group not found.
      security:
        - Session: []
  "/user/alliance/{id}/remove-group/{gid}":
    put:
      tags:
        - Alliance
      summary: Remove a group from the alliance.
      description: "Needs role: group-admin"
      operationId: removeGroup
      parameters:
        - name: id
          in: path
          description: ID of the alliance.
          required: true
          schema:
            type: integer
        - name: gid
          in: path
          description: ID of the group.
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Group removed.
        "403":
          description: Not authorized.
        "404":
          description: Alliance and/or group not found.
      security:
        - Session: []
  /user/app/all:
    get:
      tags:
        - App
      summary: List all apps.
      description: "Needs role: app-admin"
      operationId: all
      responses:
        "200":
          description: List of apps (only id and name properties are returned).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/App"
        "403":
          description: Not authorized.
      security:
        - Session: []
  /user/app/create:
    post:
      tags:
        - App
      summary: Create an app.
      description: "Needs role: app-admin<br>Generates a random secret that must be changed
        by an app manager."
      operationId: create
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  description: Name of the app.
                  type: string
                  maxLength: 255
              required:
                - name
      responses:
        "201":
          description: The new app.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "400":
          description: App name is invalid/missing.
        "403":
          description: Not authorized.
        "500":
          description: If creation of app failed.
      security:
        - Session: []
  "/user/app/{id}/rename":
    put:
      tags:
        - App
      summary: Rename an app.
      description: "Needs role: app-admin"
      operationId: rename
      parameters:
        - name: id
          in: path
          description: ID of the app.
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  description: New name for the app.
                  type: string
                  maxLength: 64
              required:
                - name
      responses:
        "200":
          description: App was renamed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "400":
          description: App name is invalid/missing.
        "403":
          description: Not authorized.
        "404":
          description: App not found.
      security:
        - Session: []
  "/user/app/{id}/delete":
    delete:
      tags:
        - App
      summary: Delete an app.
      description: "Needs role: app-admin"
      operationId: delete
      parameters:
        - name: id
          in: path
          description: ID of the app.
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: App was deleted.
        "403":
          description: Not authorized.
        "404":
          description: App not found.
      security:
        - Session: []
  "/user/app/{id}/managers":
    get:
      tags:
        - App
      summary: List all managers of an app.
      description: "Needs role: app-admin"
      operationId: managers
      parameters:
        - name: id
          in: path
          description: App ID.
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of players ordered by name. Only id, name and roles properties
            are returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
        "403":
          description: Not authorized.
        "404":
          description: App not found.
      security:
        - Session: []
  "/user/app/{id}/add-manager/{pid}":
    put:
      tags:
        - App
      summary: Assign a player as manager to an app.
      description: "Needs role: app-admin"
      operationId: addManager
      parameters:
        - name: id
          in: path
          description: ID of the app.
          required: true
          schema:
            type: integer
        - name: pid
          in: path
          description: ID of the player.
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Player added as manager.
        "403":
          description: Not authorized.
        "404":
          description: Player and/or app not found.
      security:
        - Session: []
  "/user/app/{id}/remove-manager/{pid}":
    put:
      tags:
        - App
      summary: Remove a manager (player) from an app.
      description: "Needs role: app-admin"
      operationId: removeManager
      parameters:
        - name: id
          in: path
          description: ID of the app.
          required: true
          schema:
            type: integer
        - name: pid
          in: path
          description: ID of the player.
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Player removed from managers.
        "403":
          description: Not authorized.
        "404":
          description: Player and/or app not found.
      security:
        - Session: []
  "/user/app/{id}/show":
    get:
      tags:
        - App
      summary: Shows app information.
      description: |-
        Needs role: app-admin, app-manager
             *                  Managers can only see groups of their own apps.
      operationId: show
      parameters:
        - name: id
          in: path
          description: App ID.
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The app information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "403":
          description: Not authorized.
        "404":
          description: App not found.
      security:
        - Session: []
  "/user/app/{id}/add-group/{gid}":
    put:
      tags:
        - App
      summary: Add a group to an app.
      description: "Needs role: app-admin"
      operationId: addGroup
      parameters:
        - name: id
          in: path
          description: ID of the app.
          required: true
          schema:
            type: integer
        - name: gid
          in: path
          description: ID of the group.
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Group added to app.
        "403":
          description: Not authorized.
        "404":
          description: Group and/or app not found.
      security:
        - Session: []
  "/user/app/{id}/remove-group/{gid}":
    put:
      tags:
        - App
      summary: Remove a group from an app.
      description: "Needs role: app-admin"
      operationId: removeGroup
      parameters:
        - name: id
          in: path
          description: ID of the app.
          required: true
          schema:
            type: integer
        - name: gid
          in: path
          description: ID of the group.
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Group removed from the app.
        "403":
          description: Not authorized.
        "404":
          description: Group and/or app not found.
      security:
        - Session: []
  "/user/app/{id}/add-role/{name}":
    put:
      tags:
        - App
      summary: Add a role to the app.
      description: "Needs role: app-admin"
      operationId: addRole
      parameters:
        - name: id
          in: path
          description: ID of the app.
          required: true
          schema:
            type: integer
        - name: name
          in: path
          description: Name of the role.
          required: true
          schema:
            type: string
            enum:
              - app-groups
              - app-chars
              - app-tracking
              - app-esi
      responses:
        "204":
          description: Role added.
        "403":
          description: Not authorized.
        "404":
          description: App and/or role not found or invalid.
      security:
        - Session: []
  "/user/app/{id}/remove-role/{name}":
    put:
      tags:
        - App
      summary: Remove a role from an app.
      description: "Needs role: app-admin"
      operationId: removeRole
      parameters:
        - name: id
          in: path
          description: ID of the app.
          required: true
          schema:
            type: integer
        - name: name
          in: path
          description: Name of the role.
          required: true
          schema:
            type: string
            enum:
              - app-groups
              - app-chars
              - app-tracking
              - app-esi
      responses:
        "204":
          description: Role removed.
        "403":
          description: Not authorized.
        "404":
          description: App and/or role not found or invalid.
      security:
        - Session: []
  "/user/app/{id}/change-secret":
    put:
      tags:
        - App
      summary: Generates a new application secret. The new secret is returned, it
        cannot be retrieved afterwards.
      description: "Needs role: app-manager"
      operationId: changeSecret
      parameters:
        - name: id
          in: path
          description: ID of the app.
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The new secret.
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Not authorized.
        "404":
          description: App not found.
        "500":
          description: Failed to created new secret.
      security:
        - Session: []
...
servers:
  - url: https://localhost/api
  - url: http://localhost/api
components:
  securitySchemes:
    Bearer:
      type: apiKey
      description: "Example: Bearer ABC"
      name: Authorization
      in: header
    Session:
      type: apiKey
      description: "Example: BCSESS=123"
      name: Cookie
      in: header
  schemas:
...
    App:
      required:
        - id
        - name
      properties:
        id:
          description: App ID
          type: integer
        name:
          description: App name
          type: string
          maxLength: 255
        roles:
          description: Roles for authorization.
          type: array
          items:
            $ref: "#/components/schemas/Role"
        groups:
          description: Groups the app can see.
          type: array
          items:
            $ref: "#/components/schemas/Group"
...
    Group:
      required:
        - id
        - name
      properties:
        id:
          description: Group ID.
          type: integer
        name:
          description: A unique group name (can be changed).
          type: string
          maxLength: 64
          pattern: ^[-._a-zA-Z0-9]+$
        visibility:
          type: string
          enum:
            - private
            - public
            - conditioned
...
    Role:
      type: string
      enum:
        - app
        - app-groups
        - app-chars
        - app-tracking
        - app-esi
        - user
        - user-admin
        - user-manager
        - app-admin
        - app-manager
        - group-admin
        - group-manager
        - esi
        - settings
        - tracking
...